refactor(isolate): moved definitions out of subfolder.
