--- lib/src/isolate/events/events.dart
@@ -1,2 +1,0 @@
-export 'event.dart';
-export 'request.dart';
