+++ lib/src/isolate/isolate.dart
@@ -1,0 +1,10 @@
+export 'event.dart';
+export 'isolate_dispatcher.dart';
+export 'isolate_handler.dart';
+export 'request.dart';
+
+export 'requests/build_popular_url_request.dart';
+export 'requests/chapter_request.dart';
+export 'requests/novel_request.dart';
+export 'requests/popular_request.dart';
+export 'requests/search_request.dart';
