+++ lib/src/isolate/request.dart
@@ -1,4 +1,4 @@
-import 'package:nacht_sources/src/isolate/events/events.dart';
+import 'package:nacht_sources/src/isolate/isolate.dart';
 
 /// A [Request] is an event that expects a [reply] from the isolate
 ///
@@ -8,4 +8,12 @@ class Request<T> extends Event {
 
   /// Create a reply event with this events key.
   ReplyEvent reply(T value) => ReplyEvent<T>(key, value);
+}
+
+/// Return a [value] from the isolate to the main thread.
+class ReplyEvent<T> extends Event {
+  ReplyEvent(super.key, this.value);
+
+  /// The value being sent from the isolate thread.
+  final T value;
 }
