+++ lib/src/isolate/requests/build_popular_url_request.dart
@@ -1,4 +1,4 @@
-import 'package:nacht_sources/src/isolate/events/events.dart';
+import 'package:nacht_sources/src/isolate/isolate.dart';
 
 class BuildPopularUrlRequest extends Request<String> {
   BuildPopularUrlRequest(super.key, this.page);
