+++ lib/src/isolate/requests/chapter_request.dart
@@ -1,5 +1,4 @@
-import 'package:nacht_sources/src/isolate/events/event.dart';
-import 'package:nacht_sources/src/isolate/events/events.dart';
+import 'package:nacht_sources/src/isolate/isolate.dart';
 
 class ChapterContentRequest extends Request<String?> {
   ChapterContentRequest(super.key, this.url);
