+++ lib/src/isolate/requests/novel_request.dart
@@ -1,5 +1,5 @@
 import 'package:nacht_sources/nacht_sources.dart';
-import 'package:nacht_sources/src/isolate/events/events.dart';
+import 'package:nacht_sources/src/isolate/isolate.dart';
 
 class NovelRequest extends Request<Novel> {
   NovelRequest(super.key, this.url);
