+++ lib/src/isolate/requests/popular_request.dart
@@ -1,6 +1,5 @@
 import 'package:nacht_sources/nacht_sources.dart';
-import 'package:nacht_sources/src/isolate/events/event.dart';
-import 'package:nacht_sources/src/isolate/events/events.dart';
+import 'package:nacht_sources/src/isolate/isolate.dart';
 
 class PopularRequest extends Request<List<Novel>> {
   PopularRequest(super.key, this.page);
