+++ lib/src/isolate/requests/search_request.dart
@@ -1,5 +1,5 @@
 import 'package:nacht_sources/nacht_sources.dart';
-import 'package:nacht_sources/src/isolate/events/events.dart';
+import 'package:nacht_sources/src/isolate/isolate.dart';
 
 class SearchRequest extends Request<List<Novel>> {
   SearchRequest(super.key, this.query, this.page);
